refactor: 리버스 프록시의 인증로직에서 사용하는 유틸함수 분리
